Allow reading autotracker context immediately
This change allows the autotracker context to be read as soon as the
tracker is connected, even if active gameplay is not yet started. This
should be OK because the context is written by the generator and does
not change during gameplay, and it fixes an issue where the data was not
read again when gameplay is started.